requirements.txt
Django==2.0.4
djangorestframework==3.8.2
gunicorn==19.7.1
psycopg2==2.7.4
pytz==2018.3
certifi==2018.1.18
chardet==3.0.4
idna==2.6
requests==2.18.4
urllib3==1.22
django-cors-headers==2.2.0
